Update Setup to include Python version
Wrathserver currently requires Python 3.18.13 (according to Poetry). This constraint can be changed in the pyproject.toml file, but for now the version of Python used should be shown.

Additionally, the python version used may want to be changed since it is no longer accessible through an installer and needs to be compiled from source. (This is true true through pyenv and python.org)